Just everything tbh

+++ src/main/kotlin/coffee/cypher/hexbound/feature/media_attachment/StaticMediaValue.kt
@@ -4,8 +4,8 @@ import at.petrak.hexcasting.api.misc.MediaConstants
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class StaticMediaValue(val priority: Int, val amount: Int, val unit: Unit) {
-    enum class Unit(val calculate: (Int) -> Int) {
+data class StaticMediaValue(val priority: Int, val amount: Long, val unit: Unit) {
+    enum class Unit(val calculate: (Long) -> Long) {
         ABSOLUTE({ it }),
         DUST(MediaConstants.DUST_UNIT::times),
         SHARD(MediaConstants.SHARD_UNIT::times),
